requirements.txt
# Gemini API (new SDK)
google-genai>=1.0.0

# Pinecone + embeddings for ingest step
pinecone-client==5.0.1
sentence-transformers==4.1.0

# HTTP
requests>=2.31.0

# Utilities
python-dotenv==1.0.1
